Convert to FC app (but can be started outside FC environment also). Moved plot related code to plot.py.
