+++ build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+docker build . --tag fc_visualization
