+++ requirements.txt
@@ -46,3 +46,11 @@ tqdm==4.64.0
 urllib3==1.26.9
 Werkzeug==2.0.3
 yellowbrick==1.4
+featurecloud
+bottle
+jsonpickle
+joblib
+numpy
+bios
+pydot
+pyyaml
