+++ server_config/docker-entrypoint.sh
@@ -1,0 +1,1 @@
+/usr/bin/supervisord -c "/supervisord.conf"
